requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
flask-cors==3.0.10
fuzzywuzzy==0.18.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
numpy==1.21.4
pandas==1.3.4
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-Levenshtein==0.12.2
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.26
urllib3==1.26.7
Werkzeug==2.0.2
gunicorn==20.1.0
multipledispatch==0.6.0
